% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequencebox.R
\name{create_tree}
\alias{create_tree}
\title{Create tree}
\usage{
create_tree(seqs)
}
\arguments{
\item{seqs}{Sequences}
}
\description{
Create a phylogenetic tree
}
\examples{
ref <- generate_random_sequence("ATGC", 10)
seq <- random_sequence_mod(ref, 5)
align_result <- align(ref, seq, mode = "common")
create_tree(align_result$sequences) 

}
