% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequencebox.R
\name{generate_random_sequence}
\alias{generate_random_sequence}
\title{Generate random sequence}
\usage{
generate_random_sequence(src, len)
}
\arguments{
\item{src}{Source string to get characters out of}

\item{len}{Length of the random sequences}
}
\description{
Generate random sequence
}
\examples{
generate_random_sequence("ATGC", 10)

}
