% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequencebox.R
\name{get_sequence_diffs}
\alias{get_sequence_diffs}
\title{Get sequence differences}
\usage{
get_sequence_diffs(reference, sequences)
}
\arguments{
\item{reference}{Reference sequence or an array of reference sequences. Assumed aligned.}

\item{sequences}{An array of sequences to be compared to the reference}
}
\description{
For sequences ABCD and AXCY the differences would be B2X and D4Y
}
\examples{
ref <- generate_random_sequence("ATGC", 10)
seqs <- random_sequence_mod(ref, 10)
aligned <- align_sequences(ref, seqs, mode = "common")
get_sequence_diffs(aligned$references, aligned$sequences)

}
