% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequencebox.R
\name{plot_alignment_matrix}
\alias{plot_alignment_matrix}
\title{Plot alignment matrix}
\usage{
plot_alignment_matrix(score, direction, ref, seq)
}
\arguments{
\item{score}{Matrix with scores}

\item{direction}{Matrix with directions}

\item{ref}{Reference sequence}

\item{seq}{Sequence that was aligned}
}
\description{
Plot a Needleman–Wunsch alignment matrix. Consists of 2 parts -
a score matrix and a direction matrix.
}
\examples{
ref <- "ABCD"
seq <- "ACC"
align_result <- align(ref, seq, matrices = TRUE)
plot_alignment_matrix(align_result$scores[[1]], align_result$directions[[1]], ref, seq)

}
