% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequencebox.R
\name{plot_sequences}
\alias{plot_sequences}
\title{Plot sequences}
\usage{
plot_sequences(reference, sequences)
}
\arguments{
\item{reference}{Reference sequence}

\item{sequences}{An array of sequences to be plotted}
}
\description{
Plot a list of sequences coloring each letter accoring to the reference.
}
\examples{
plot_sequences("ABCDEFGH", c("AAACEF", "ABCDEA"))

}
