% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequencebox.R
\name{plot_tree}
\alias{plot_tree}
\title{Plot tree}
\usage{
plot_tree(
  tree,
  root = 0,
  tipnames = paste0("tip", tree$node[!tree$node$internal, "node"])
)
}
\arguments{
\item{tree}{Tree}

\item{root}{Root node index}
}
\description{
Plot a phylogenetic tree
}
\examples{
ref <- generate_random_sequence("ATGC", 10)
seq <- random_sequence_mod(ref, 5)
align_result <- align(ref, seq, mode = "common")
tree <- create_tree(align_result$sequences)
plot_tree(tree)
}
