% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequencebox.R
\name{random_sequence_mod}
\alias{random_sequence_mod}
\title{Generate random sequence}
\usage{
random_sequence_mod(
  src,
  trim_start_max = 0.1,
  trim_end_max = 0.1,
  mutation = 0.1,
  deletion = 0.1,
  insertion = 0.1,
  insertion_src = "ATGC"
)
}
\arguments{
\item{src}{Source string to get characters out of}

\item{len}{Length of the random sequences}
}
\description{
Generate random sequence
}
\examples{
seq <- generate_random_sequence("ATGC", 30)
print(seq)
random_sequence_mod(seq)

}
