% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequencebox.R
\name{align}
\alias{align}
\title{Align sequences}
\usage{
align(reference, sequences)
}
\arguments{
\item{reference}{Reference sequence}

\item{sequences}{An array of sequences to be plotted}
}
\description{
Align a list of sequences to a reference
}
\examples{
align("ABCDEFGH", c("AAACEF", "ABCDEA"))

}
