% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequencebox.R
\name{align}
\alias{align}
\title{Align sequences}
\usage{
align(reference, sequences, mode = "individual", matrices = FALSE)
}
\arguments{
\item{reference}{Reference sequence or an array of reference sequences (array is only allowed if mode is 'individual')}

\item{sequences}{An array of sequences to be aligned to the reference(s)}

\item{mode}{'individual' to align strings to reference(s) individually or 'common' to align to one common reference.}

\item{matrices}{Return alignment matrices}
}
\description{
Align a list of sequences to a reference
}
\examples{
ref <- generate_random_sequence("ATGC", 10)
seqs <- random_sequence_mod(ref, 10)
align(ref, seqs, mode = "common")

}
