% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequencebox.R
\name{plot_sequences}
\alias{plot_sequences}
\title{Plot sequences}
\usage{
plot_sequences(reference, sequences)
}
\arguments{
\item{reference}{Reference sequence}

\item{sequences}{An array of sequences to be plotted}
}
\description{
Plot a list of sequences coloring each letter accoring to the reference.
}
\examples{
ref <- generate_random_sequence("ATGC", 10)
seqs <- random_sequence_mod(ref, 10)
seqs_aligned <- align_sequences(ref, seqs, mode = "common")
plot_sequences(seqs_aligned$references, seqs_aligned$sequences)

}
